Don't remember making this change
But the git dif says it's there and I trust past me more than current me

+++ src/main/java/net/tazgirl/armour_rebalance/mixins/LivingEntityMixin.java
@@ -83,7 +83,7 @@ public abstract class LivingEntityMixin
         DamageContainer damageContainer = damageContainers.peek();
         LivingEntity me = (LivingEntity)(Object) this;
 
-        DamageReduction.applyFullReduction(damageContainer, me, damageSource, damageAmount, (float) me.getAttributeValue(Attributes.ARMOR),(float) me.getAttributeValue(Attributes.ARMOR_TOUGHNESS), true, true, true, true);
+        DamageReduction.applyFullReduction(damageContainer, me, damageSource, damageAmount, me.getAttributeValue(Attributes.ARMOR),me.getAttributeValue(Attributes.ARMOR_TOUGHNESS), true, true, true, true);
     }
 }
